move user authentication pages to their own package
