+++ app/src/main/AndroidManifest.xml
@@ -15,20 +15,20 @@
         android:theme="@style/Theme.Planetze35"
         tools:targetApi="31">
         <activity
-            android:name=".EmailVerificationActivity"
+            android:name=".userauthentication.EmailVerificationActivity"
             android:exported="false"
             android:label="Verify you email" />
         <activity
-            android:name=".ForgotPasswordActivityView"
+            android:name=".userauthentication.ForgotPasswordActivityView"
             android:exported="false"
-            android:parentActivityName=".LoginActivityView" />
+            android:parentActivityName=".userauthentication.LoginActivityView" />
         <activity
-            android:name=".SignupActivity"
+            android:name=".userauthentication.SignupActivity"
             android:exported="false"
-            android:parentActivityName=".LoginActivityView"
+            android:parentActivityName=".userauthentication.LoginActivityView"
             android:label="Planetze35 - signup" />
         <activity
-            android:name=".LoginActivityView"
+            android:name=".userauthentication.LoginActivityView"
             android:exported="true"
             android:theme="@style/Theme.Planetze35.Actionbar">
             <intent-filter>
