+++ app/src/main/java/com/example/planetze35/userauthentication/Contract.java
@@ -1,4 +1,4 @@
-package com.example.planetze35;
+package com.example.planetze35.userauthentication;
 
 public interface Contract {
