+++ app/src/main/java/com/example/planetze35/userauthentication/EmailUtils.java
@@ -1,4 +1,4 @@
-package com.example.planetze35;
+package com.example.planetze35.userauthentication;
 
 import android.content.Context;
 import android.widget.Toast;
