+++ app/src/main/java/com/example/planetze35/userauthentication/LoginModel.java
@@ -1,4 +1,4 @@
-package com.example.planetze35;
+package com.example.planetze35.userauthentication;
 
 import android.graphics.Color;
 import android.widget.Toast;
