+++ app/src/main/res/layout/activity_email_verification.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EmailVerificationActivity">
+    tools:context=".userauthentication.EmailVerificationActivity">
 
     <TextView
         android:id="@+id/tvEmailVerificationMessage"
