+++ app/src/main/res/layout/activity_forgot_password.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ForgotPasswordActivityView">
+    tools:context=".userauthentication.ForgotPasswordActivityView">
 
     <Button
         android:id="@+id/btnForgotPasswordEmail"
