+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:id = "@+id/main"
     android:layout_width = "match_parent"
     android:layout_height = "match_parent"
-    tools:context = ".LoginActivityView">
+    tools:context = ".userauthentication.LoginActivityView">
 
     <TextView
         android:id = "@+id/tvTitleLoginPage"
